pyproject.toml
[tool.poetry]
name = "event-management-system"
version = "0.1.0"
description = ""
authors = ["MaximPyanin <maximpyanin@icloud.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
pylint = "^3.2.2"
black = "^24.4.2"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
bcrypt = "^4.1.3"
fastapi = "^0.111.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
cryptography = "^42.0.8"
httpx = "^0.27.0"
dependency-injector = "^4.41.0"
gunicorn = "^22.0.0"
uvicorn = "^0.30.1"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
sendgrid = "^6.11.0"
twilio = "^9.2.1"
apscheduler = "^3.10.4"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
